refactor: 디렉토리 리펙토링
