+++ src/main/java/lotto/domain/BonusComponent.java
@@ -1,4 +1,4 @@
-package lotto.domain.lotto;
+package lotto.domain;
 
 public interface BonusComponent {
 }
