+++ src/main/java/lotto/domain/Lottos.java
@@ -1,6 +1,6 @@
-package lotto.domain.lotto;
+package lotto.domain;
 
-import lotto.domain.lotto.factory.LottoFactory;
+import lotto.domain.factory.LottoFactory;
 
 import java.util.ArrayList;
 import java.util.List;
